Add batch inference support for Qwen3-VL (Python + Java)
- Implement batched decoder: vision encoder per-image, decoder batched
- Qwen3VlModel: lastImageResult -> lastImageResults list, batch embed/position/deepstack/decode
- Fix Qwen3VlEncoderModel: long-to-int cast for numDeepstack
- Add Python batch inference script (qwen3vl_batch_onnx_infer.py)
- Add Java batch test (Qwen3VlBatchTest.java)
- Validated: batch==sequential for both Python and Java

+++ flux-ocr/src/main/java/io/github/flux/qwen3vl/Qwen3VlEncoderModel.java
@@ -40,7 +40,7 @@ public class Qwen3VlEncoderModel implements AutoCloseable {
             }
             this.session = env.createSession(modelFile, options);
             // Detect deepstack count from ONNX model outputs: total outputs - 1 (image_features)
-            this.numDeepstack = session.getNumOutputs() - 1;
+            this.numDeepstack = (int) (session.getNumOutputs() - 1);
         } catch (Exception e) {
             throw new FluxException(e);
         }
